pyproject.toml
[tool.poetry]
name = "online-shop-vertica-dwh"
version = "0.1.0"
description = ""
authors = ["Leonidee"]
readme = "README.md"
packages = [{include = "online_shop_vertica_dwh"}]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26.104"
coloredlogs = "^15.0.1"
pandas = "^1.5.3"
tabulate = "^0.9.0"
faker = "^18.3.2"
python-dotenv = "^1.0.0"
pyarrow = "^11.0.0"
vertica-python = "^1.3.1"
ipykernel = "^6.22.0"
apache-airflow = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
